+++ examples/contact/find.php
@@ -37,8 +37,12 @@ $response_messages = $response->ResponseMessages->FindItemResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to search for contacts with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(
+            STDERR,
+            "Failed to search for contacts with \"$code: $message\"\n"
+        );
         continue;
     }