+++ examples/event/find.php
@@ -51,8 +51,12 @@ $response_messages = $response->ResponseMessages->FindItemResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to search for events with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(
+            STDERR,
+            "Failed to search for events with \"$code: $message\"\n"
+        );
         continue;
     }