+++ examples/event/sync.php
@@ -55,11 +55,9 @@ $response_messages = $response->ResponseMessages
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to sync folder with \"$message\"\n");
-
-        $soap = $client->getClient();
-        $lr = $soap->__getLastRequest();
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Failed to sync folder with \"$code: $message\"\n");
         continue;
     }