+++ examples/event/update.php
@@ -80,8 +80,9 @@ $response_messages = $response->ResponseMessages->UpdateItemResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to update event with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Failed to update event with \"$code: $message\"\n");
         continue;
     }