+++ examples/folder/create.php
@@ -53,8 +53,9 @@ $response_messages = $response->ResponseMessages->CreateFolderResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Folder failed to create with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Folder failed to create with \"$code: $message\"\n");
         continue;
     }