+++ examples/message/add-attachment.php
@@ -55,8 +55,12 @@ $response_messages = $response->ResponseMessages
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to create attachment with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(
+            STDERR,
+            "Failed to create attachment with \"$code: $message\"\n"
+        );
         continue;
     }