+++ examples/message/create.php
@@ -78,8 +78,9 @@ $response_messages = $response->ResponseMessages->CreateItemResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Message failed to create with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Message failed to create with \"$code: $message\"\n");
         continue;
     }