+++ examples/message/describe.php
@@ -59,8 +59,9 @@ $response_messages = $response->ResponseMessages->GetItemResponseMessage;
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to get message with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Failed to get message with \"$code: $message\"\n");
         continue;
     }