+++ examples/system/timezone-definitions.php
@@ -40,8 +40,9 @@ $response_messages = $response->ResponseMessages
 foreach ($response_messages as $response_message) {
     // Make sure the request succeeded.
     if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-        $message = $response_message->ResponseCode;
-        fwrite(STDERR, "Failed to get timezones with \"$message\"\n");
+        $code = $response_message->ResponseCode;
+        $message = $response_message->MessageText;
+        fwrite(STDERR, "Failed to get timezones with \"$code: $message\"\n");
         continue;
     }