+++ examples/user/meeting-suggestions.php
@@ -72,8 +72,12 @@ $response = $client->GetUserAvailability($request);
 // Make sure the request succeeded.
 $response_message = $response->SuggestionsResponse->ResponseMessage;
 if ($response_message->ResponseClass != ResponseClassType::SUCCESS) {
-    $message = $response_message->ResponseCode;
-    fwrite(STDERR, "Failed to get user availability with \"$message\"\n");
+    $code = $response_message->ResponseCode;
+    $message = $response_message->MessageText;
+    fwrite(
+        STDERR,
+        "Failed to get user availability with \"$code: $message\"\n"
+    );
     exit;
 }