+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed Exchange 2007 version parsing in autodiscovery.
 - Fixed event title in event/create example.
 - Fixed capitalization of URL property on PushSubscriptionRequestType.
+- Improved error messages in examples.
 
 ## 1.0.0-beta.2 - 2017-01-12